Split income and outgoings to two seperate classes, must be called individually from main menu now
